Move rendering code into module
