+++ game.c
@@ -1,100 +1,17 @@
 #include <time.h>
 #include <stdlib.h>
-#include <ncurses.h>
 #include <string.h>
 
 #include "mob/mob.h"
 #include "level/level.h"
 
+#include "curses_renderer.h"
 #include "game.h"
 #include "log.h"
 #include "input.h"
 #include "simulation/simulation.h"
 #include "los/los.h"
 #include "colors/colors.h"
-
-
-int keyboard_x = 0, keyboard_y = 0;
-char message_banner[MESSAGE_LENGTH];
-
-void draw_mobile(mobile *mob, int x_offset, int y_offset) {
-    char icon = ((item*)mob)->display;
-
-    if (mob->emote) {
-        icon = mob->emote;
-        mob->emote = false;
-    }
-
-    mvprintw(mob->y + y_offset, mob->x + x_offset, "%c", icon);
-}
-
-void draw(level *lvl) {
-    int row,col;
-    getmaxyx(stdscr,row,col);
-    row -= 1;
-
-    // Offset to keep player in center
-    int x_offset = col / 2 - lvl->player->x;
-    int y_offset = row / 2 - lvl->player->y;
-
-    // Draw map and items
-    for (int xx = 0; xx < col; xx++) {
-        for (int yy = 0; yy < row; yy++) {
-            // (xx,yy) are screen coordinates
-            // (x ,y ) are map coordinates
-            int x = xx - x_offset;
-            int y = yy - y_offset;
-
-            char icon = UNSEEN;
-
-            if ((0 <= x && x < lvl->width) && (0 <= y && y < lvl->height)) {
-                //TODO wrapper function with clear name
-                if (can_see(lvl, lvl->player, x, y)) {
-                    if (lvl->chemistry[x][y]->elements[fire] > 0) {
-                        icon = BURNING;
-                        attron(COLOR_PAIR(RED));
-                    } else if (lvl->items[x][y] != NULL) {
-                        icon = lvl->items[x][y]->item->display;
-                    } else {
-                        icon = lvl->tiles[x][y];
-                    }
-                }
-                // Fog of war
-                if (icon == UNSEEN) {
-                    icon = lvl->memory[x][y];
-                    attron(COLOR_PAIR(YELLOW));
-                } else {
-                    lvl->memory[x][y] = icon;
-                }
-            }
-            mvprintw(yy, xx, "%c", icon);
-            //TODO this should be a general unsetter, not this
-            attroff(COLOR_PAIR(YELLOW));
-            attroff(COLOR_PAIR(RED));
-        }
-    }
-
-    // Draw mobs
-    for (int i=0; i < lvl->mob_count; i++) {
-        mobile* mob = lvl->mobs[i];
-        if (mob->active && can_see(lvl, lvl->player, mob->x, mob->y)) {
-            if ((0 < mob->y + y_offset && mob->y + y_offset <= row) && (0 < mob->x + x_offset && mob->x + x_offset <= col)) {
-                draw_mobile(mob, x_offset, y_offset);
-            }
-        }
-    }
-    draw_mobile(lvl->player, x_offset, y_offset);
-    move(row, 0);
-    clrtoeol();
-    mvprintw(row, 0, message_banner);
-}
-
-//TODO When the same message (e.g. quaff) is repeated, it
-//should be clear somehow that there were multiple messages
-void print_message(char *msg) {
-    strncpy(message_banner, msg, MESSAGE_LENGTH);
-}
-
 
 
 void step_chemistry(chemical_system *sys, constituents *chem, constituents *context) {
@@ -222,23 +139,11 @@ int main() {
         if (do_log != NULL) logging_active = true;
 
         srand(time(NULL));
-        initscr();
 
-        if (! init_colors()) {
-            //TODO print TERM environment variable
-            logger("Terminal does not support color.\n");
-            exit(1);
-        }
-
-        raw();
-        noecho();
-        cbreak();
-        keypad(stdscr, TRUE);
-
-        curs_set(0);
+        init_rendering_system();
 
         lvl = make_level();
-        draw(lvl);
+        draw_level(lvl);
 
         // Main Loop
         while (lvl->active) {
@@ -261,11 +166,11 @@ int main() {
             }
 
             // Update the screen
-            draw(lvl);
+            draw_level(lvl);
 
             // If the player is dead, wait for input
             if (!lvl->player->active) {
-                getch();
+                get_input(lvl);
                 break;
             }
 
@@ -276,6 +181,7 @@ int main() {
         }
 
         destroy_level(lvl);
-        endwin();
+
+        cleanup_rendering_system();
         return 0;
 }
